Create pump.cpp
A new, separate source file has been created for the pump implementation. This uses a class structure rather than functions in a namespace. This change was made so that the pump functions would have access to the pin number for controlling the pump without having to make the pin number a global variable.