Add documentation to PumpCommand
Added a brief summary of the three PumpCommand classes (PumpOn,
PumpOff, and PumpOnTimed). The documentation explains the
difference between this implementation of the command pattern and
the standard Gang of Four version.

An example usage of how one of the PumpCommands can be used. The
example uses PumpOnTimed to turn a pump on for five minutes. A
reminder is also added about the need to regularly call the pump's
update() during timed runs to ensure that the pump shuts off at
the right time.