Add documentation for pump
Added a brief description of the Pump class and how to construct
it. Included a link to a webpage where more details on the
Raspberry Pi pin numbering scheme can be found.

Added an example of creating and using a Pump object to run a
pump that is connected to GPIO pin 17 for 30 seconds. Added a
reminder about the need to regularly call the update() method on
the pump during timed pump runs.