Implement thread safe function in PumpCommandQueue
The PumpCommandQueue was always intended to be thread safe.
Thread safe behaviour has now been implemented through the use of
std::mutex and std::condition_variable.

In addition, out of date comments have been removed from the
methods of PumpCommandQueue, and comments documenting the
functionality of the now thread safe methods have been added.

The example provided in the PumpCommandQueue documentaion has
been updated. Thisd update reflects a change to the recommend way
of creating the command queue. The previous recommendation was to
create the queue as a std::shared_ptr and then pass that pointer
to the PumpManager class. However, the PumpManager class no longer
exists; it has been replaced by the function PumpManagementTask.
Whereas the previous PumpManager.cont_task took a
std::shared_ptr<PumpCommandQueue> as an input parameter for the
queue, the PumpManagementTask takes a reference to
PumpCommandQueue (PumpCommandQueue&).