Update main.cpp to use new pump functionality
The way in which the main function uses the classes and
functions provided by the pump.h header are updated. The
interface to the Pump has changed since the last time main.cpp
was updated, and now main.cpp is updated to use the new
interface.

The creation of the Pump now happens within the main() function
and is no longer a global variable. Also, the pump is now created
as a std::shared_ptr<Pump>, as is required by the interface of
the PumpManagementTask.

The PumpCommandQueue is no longer created as std::shared_ptr,
creation of pump_command_queue is now as follows.

	PumpCommandQueue pump_command_queue;

The exit_signal for the PumpManagementTask has been created in
main as a std::shared_ptr<bool>. This is required by the interface
of PumpManagementTask.

The PumpManagementTask has been set up to run in its own thread:

	thread pump_management_thread(std::bind(PumpManagementTask, std::ref(pump_command_queue), pump_ptr, exit_signal));

The main.cpp function querry_handler has been modified to take a
reference to PumpCommandQueue (PumpCommandQueue&) instead of a
std::shared_ptr<PumpCommandQueue>. The implementation of
querry_handler has been changed to relfect that pump_command_queue
is a reference, rather than a pointer. (i.e., "->" operator has
been changed to "." for all pump_command_queue function calls).